server/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
Werkzeug
flask-cors
flask-jwt-extended
python-dotenv
requests